requirements.txt
pytest
allure-pytest
requests
flake8
black